table tests: expand them considerably and actually run them
In the current version, the runTimeMapTest binary was run because of a
copy-and-pasto. oops.

Also, not all of the tests pass but that is because of another problem
of opm-parser for which I don't know a good fix: It currently seems to
be impossible to specify empty records if there are non-defaulted
items in the JSON keyword specification, but such empty records are
used as table separators by Eclipse...