GridProperties: don't use the garbage from defaulted data items
e.g.

```
SWL
* /
```

resulted in the SWL keyword exhibiting in a single item of size 1 with
0 (or possibly undefined) value. without the previous patch, there is
not even a way to detect of that item was defaulted or not.

Note that the approach here is still incomplete as it can't support
situations like

```
SWL
10*0.1 10* 10*0.2 /
```

to fix this properly, more fundamental changes to the opm-parser data
structures are required, though...