get rid of the 'strict parsing' concept
the 'parsing' stage (i.e., the creation of the deck object) is now
always as relaxed as possible, but warnings are added to the log
object. if you want to verify more aspects of the validity of a deck,
use Opm::checkDeck(deck, parserLog). after this,
parserLog->getMessage() and friends can be used for finding out the
details what is wrong!