Rudimentary example of Log singleton.
This commit shows a very basic example of how Logging singleton can be
implemented. The commit consists of a fully static class OpmLog which
manages a Logger class; in this case an instance of the ParserLog
class. The static OpmLog class can then be used from anywhere in the opm
code with:

  OpmLog::addMessage( messageType , "Log message")

Before this can be really useful the Logger class should be extended /
subclassed / ... for general usability.