add the keywords required to specify the initial reservoir temperature
E100 does not know about TEMPI, but it is wise not to use the
geothermal gradient directly, but to go over the "temperature at
cell-center" detour as specified by the E300 TEMPI keyword. The depth
vs. temperature table is given by the RTEMPVD keyword. The intend is
that if TEMPI stays unspecified the temperature is calculated at each
cell center and the result is put into TEMPI from where it gets picked
up by the simulator.

(Note that E300 says that TEMPVD is an alias for RTEMPVD even though
the possible sections of these two do not exhibit any overlap, WTF?)