add pseudo default values for more items
this is required by the Norne deck. the EQLOPTS keyword seems weird,
though, because its records seem to have a variable number of items...

for most of these the RM says that if nothing is specified, a value of
another item should be used. some items even have legitimate
defaults, but they are only mentioned in the prose of the RM's item
description. (e.g.the wetting_phase_flag item of the EHYSTR keyword...)