make the 'd' exponent specifier the slow path
i.e. first try to convert it normally and only if this fails, replace
'd' by 'e' and try again.

this is because the slowdown when always taking the second path was
about 7 to 9% for the Norne deck on my machine. (I think this is quite
surprising.)

thanks to [at]joakim-hove for the hint.