DeckKeyword begin/end uses Utility/Iterator
Refactors previous begin begin/end support to not expose the raw
vector's iterator, but rather use the new Utility/Iterator.  This makes
the iterator behave better (i.e. more like std::iterators) giving access
to the raw data, not the implementation detail of the pointer to it.

This breaks all previous uses of begin/end which have also been updated.