Increase warning level by default in debug mode
When compiling in debug mode, set a series of previously unused
warnings by changing a dead code path in the cmake file. For release
builds warnings aren't as interesting and is largely turned off.

To keep the user in control of exactly what warnings are used, the
CMAKE_CXX_FLAGS are expaneded *after* the defaults are set, so that the
Wno- family of options can be used to override the defaults, rather than
just extending the defaults.