Privatise Deck; don't expose Deck via Parser.hpp
Removes the exposing of Deck from Parser.hpp. In order to obtain a Deck
instance, the new functions parseDeck and parseDeckString, placed in the
ecl namespace in bits/Parsers.hpp can be used. They are essentially
renames of the static Parser::parseFile and Parser::parseString. This
also mandates the move of Parser::parse and Parser::parseGrid which
produces EclipseState and EclipseGrid instances respectively. These
functions are moved to the ecl namespace in EclipseState.hpp

This introduces a *huge* set of broken tests, which are really
search-and-replace.