REFs hack for opm-material
This is to facilitate the privatisation of the Deck object and
redefinition of the opm-parser interfaces. The opm-material module reads
the deck directly to extract temperature information, but this is
largely unsupported in opm-parser.

In order to not break downstream functionality we take ownership of
the opm-material implemented parsing of THERMX1 and friends and expose
their raw values via the REFs and get_refs_material function. This is
ONLY meant for compatibility and SHOULD NOT BE CALLED in new code.