make the travis CI scripts independent of opm-common
the reason is that everything in opm-parser is now independent of
opm-common, so why should the travis support scripts still rely on it?

besides this, the downstreams are not build and tested by travis
anymore because travis recently aquired the habit of timing out if
this is enabled, so the downstream builder is more of a distraction
than a help. also, the OPM downstream modules nowadays ought to use
the OPM Jenkins rig ( http://ci.opm-project.org ) as their primary CI
system anyway.