make the build system compatible with Dune's cmake build system
before this, dune modules that are not using the OPM build system
needed to treat opm-parser as an external library. With this patch,
opm-parser can be build as part of duncontrol build chains, i.e.,
other modules just need to specify a depencency on opm-parser in their
dune.module file and opm-parser's libraries and headers get picked up
just fine.

Things like generating opm-parser-config.cmake should quite likely be
done in a better way, but I'm not really a cmake expert and the
solution proposed in this patch seems to get things done regardless of
whether opm-parser is system-installed or not. Comprehensive solutions
are highly appreciated.